Stabilize project: remove Lombok, fix null safety, and update security config

+++ src/main/java/com/example/website_quan_li_hoat_dong_tu_thien/repository/BlogRepository.java
@@ -4,5 +4,5 @@ import com.example.website_quan_li_hoat_dong_tu_thien.model.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface BlogRepository extends JpaRepository<Blog, Long> {
+public interface BlogRepository extends JpaRepository<Blog, Integer> {
 }
