+++ src/main/java/com/example/website_quan_li_hoat_dong_tu_thien/repository/SlideRepository.java
@@ -4,5 +4,5 @@ import com.example.website_quan_li_hoat_dong_tu_thien.model.Slide;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface SlideRepository extends JpaRepository<Slide, Long> {
+public interface SlideRepository extends JpaRepository<Slide, Integer> {
 }
